Apply same query pattern to Requests
