+++ atst/domain/common/__init__.py
@@ -1,0 +1,1 @@
+from .query import Query
